requirements.txt
asgiref==3.2.3
boto3==1.11.2
botocore==1.14.2
dj-database-url==0.5.0
Django==3.0.2
django-environ==0.4.5
django-filter==2.2.0
django-heroku==0.3.1
django-rest-auth==0.9.5
django-storages==1.8
djangorestframework==3.11.0
docutils==0.15.2
jmespath==0.9.4
Markdown==3.1.1
psycopg2==2.8.4
python-dateutil==2.8.1
pytz==2019.3
s3transfer==0.3.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
whitenoise==5.0.1
